+++ enlighten/device/AccessoryControlXSFeature.py
@@ -3,6 +3,8 @@ import logging
 from enlighten.ui.ScrollStealFilter import ScrollStealFilter
 from enlighten.EnlightenFeature import EnlightenFeature
 
+from wasatch.AccessoryConnector import AccessoryConnector
+
 log = logging.getLogger(__name__)
 
 class AccessoryControlXSFeature(EnlightenFeature):
@@ -20,8 +22,7 @@ class AccessoryControlXSFeature(EnlightenFeature):
     | | Direction: [_INPUT____v] | |
     | | Function:  [_DISABLED_v] | | (..., EXT_TRIGGER_RISING_EDGE, LASER_OVERRIDE)
     | | Value:     1 (HIGH)      | |
-    | |__________________________| |
-    |                              |
+    | |__________________________| | |                              |
     | .-[ GPIO2 ]----------------. |
     | | Mode:      [_MANUAL___v] | |
     | | Direction: [_OUTPUT___v] | |
@@ -49,7 +50,6 @@ class AccessoryControlXSFeature(EnlightenFeature):
         self.frame_cont_strobe      = cfu.frame_xs_acc_cont_strobe
         self.frame_gpio1            = cfu.frame_xs_acc_gpio1
         self.frame_gpio2            = cfu.frame_xs_acc_gpio2
-
         self.cb_gpio_enable         = cfu.checkBox_xs_acc_gpio_enable
         self.cb_acc_5v_enable       = cfu.checkBox_xs_acc_5V_enable
         self.combo_gpio1_mode       = cfu.comboBox_xs_acc_gpio1_mode
@@ -65,9 +65,8 @@ class AccessoryControlXSFeature(EnlightenFeature):
         self.sb_strobe_delay_us     = cfu.spinBox_xs_acc_cont_strobe_delay_us
         self.sb_strobe_count        = cfu.spinBox_xs_acc_cont_strobe_count
 
-
         self.visible = False
-        self.gpio_enabled = False
+        self.acc_gpio_enabled = False
         self.acc_5v_enabled = False
         self.cont_strobe_enabled = False
 
@@ -85,9 +84,37 @@ class AccessoryControlXSFeature(EnlightenFeature):
         self.update_visibility()
 
     def update_settings(self):
-        self.gpio_enabled = self.cb_gpio_enable.isChecked()
+        self.acc_gpio_enabled = self.cb_gpio_enable.isChecked()
         self.acc_5v_enabled = self.cb_acc_5v_enable.isChecked()
-        self.cont_strobe_enabled = self.gpio_enabled and self.combo_gpio2_func.currentText() == "Cont Strobe"
+        self.cont_strobe_enabled = self.acc_gpio_enabled and self.combo_gpio2_func.currentIndex() == AccessoryConnector.GPIO2_FUNCTION_CONT_STROBE
+
+        spec = self.ctl.multispec.current_spectrometer()
+        if spec:
+            acc = spec.settings.state.accessory_connector
+            if acc:
+                acc.acc_gpio_enabled = self.acc_gpio_enabled
+                acc.acc_5v_enabled = self.acc_5v_enabled
+
+                acc.state_gpio1.mode = self.combo_gpio1_mode.currentIndex()
+                acc.state_gpio1.dir = self.combo_gpio1_dir.currentIndex()
+                if acc.state_gpio1.mode == AccessoryConnector.MODE_FUNCTION:
+                    acc.state_gpio1.function = self.combo_gpio1_func.currentIndex()
+                if acc.state_gpio1.dir == AccessoryConnector.DIR_OUTPUT:
+                    acc.state_gpio1.value = self.combo_gpio1_value.currentIndex()
+
+                acc.state_gpio2.mode = self.combo_gpio2_mode.currentIndex()
+                acc.state_gpio2.dir = self.combo_gpio2_dir.currentIndex()
+                if acc.state_gpio2.mode == AccessoryConnector.MODE_FUNCTION:
+                    acc.state_gpio2.function = self.combo_gpio2_func.currentIndex()
+                if acc.state_gpio2.dir == AccessoryConnector.DIR_OUTPUT:
+                    acc.state_gpio2.value = self.combo_gpio2_value.currentIndex()
+
+                acc.strobe_period_us = self.sb_strobe_period_us.value()
+                acc.strobe_width_us = self.sb_strobe_width_us.value()
+                acc.strobe_delay_us = self.sb_strobe_delay_us.value()
+                acc.strobe_count = self.sb_strobe_count.value()
+
+                spec.change_device_setting("sync_acc")
 
         self.update_visibility()
 
@@ -96,8 +123,14 @@ class AccessoryControlXSFeature(EnlightenFeature):
         if spec is None:
             return
         
-        spec.change_device_setting("acc_5v_enable", False)
-        spec.change_device_setting("acc_gpio_enable", False)
+        acc = spec.settings.state.accessory_connector
+        if acc is None:
+            return
+
+        # might want some logic around these
+        acc.acc_gpio_enabled = False
+        acc.acc_5v_enabled = False
+        spec.change_device_setting("sync_acc")
 
     def update_visibility(self):
         spec = self.ctl.multispec.current_spectrometer()
@@ -107,7 +140,7 @@ class AccessoryControlXSFeature(EnlightenFeature):
 
         for w in [ self.lb_gpio1, self.frame_gpio1,
                    self.lb_gpio2, self.frame_gpio2 ]:
-            w.setVisible(self.gpio_enabled)
+            w.setVisible(self.acc_gpio_enabled)
 
         for w in [ self.lb_cont_strobe, self.frame_cont_strobe ]:
             w.setVisible(self.cont_strobe_enabled)