+++ enlighten/device/BatteryFeature.py
@@ -1,6 +1,4 @@
 import logging
-import datetime
-import pyqtgraph
 
 from enlighten.EnlightenFeature import EnlightenFeature
 
@@ -15,22 +13,14 @@ class BatteryFeature(EnlightenFeature):
 
         cfu = ctl.form.ui
 
-        self.curve                 = None
-        self.name                  = "Battery"
-        self.output_to_file        = False
+        self.lb_hw_view_level = cfu.label_hw_view_battery_parsed
+        self.lb_power_connection_state = cfu.label_hw_view_power_connection_state
 
-        self.lb_perc = cfu.label_hardware_capture_details_battery
-        self.lb_parsed = cfu.label_battery_parsed
+        self.strip_chart_level = self.ctl.strip_charts.create_chart(name="Battery Charge Level", y_unit="Percent (%)", fmt="%.2f%%", warn_lo=0.20)
 
-        # MZ: consider eventually moving this to a new Feature, but fine for now
-        self.lb_power_connection_state = cfu.label_power_connection_state
-
-        self.populate_placeholder()
-        ctl.multispec.register_strip_feature(self)
-        ctl.hardware_file_manager.register_feature(self)
-
-        cfu.pushButton_battery_clear_history.clicked.connect(self.clear_data)
-        cfu.pushButton_battery_copy_history.clicked.connect(self.copy_data) 
+        # todo: add strip_charts for:
+        # - reading.battery_temperature_deg_c
+        # - reading.battery_charger_temperature_deg_c
 
     def process_reading(self, spec, reading):
         if reading.power_connection_state:
@@ -39,128 +29,26 @@ class BatteryFeature(EnlightenFeature):
         if reading.battery_percentage is None:
             return
 
-        if hasattr(spec.settings.eeprom,"has_battery"):
-            if not spec.settings.eeprom.has_battery:
-                self.lb_perc.setText("No Battery")
-                return
+        if not spec.settings.eeprom.has_battery:
+            return
     
-        # add state.battery_temperature_deg_c
-        # add state.battery_charger_temperature_deg_c
-
         self.process_battery_charge_level(spec, reading)
 
     def process_battery_charge_level(self, spec, reading):
-        app_state = spec.app_state
-        rds = app_state.battery_data
-
         perc = reading.battery_percentage
         is_charging = reading.battery_charging
         charging_label = 'charging' if is_charging else 'discharging'
 
-        log.debug(f"adding {perc} to RDS")
-        rds.add(perc)
-        current_time = datetime.datetime.now()
-        if self.output_to_file:
-            self.ctl.hardware_file_manager.write_line(self.name, f"{self.name}, {spec.label}, {current_time}, {perc}, {charging_label}")
+        self.strip_chart_level.add_value(perc, spec=spec)
 
-        self.lb_parsed.setText(f"{perc:.2f}%, {charging_label}")
-        if spec == self.ctl.multispec.current_spectrometer():
-            self.lb_perc.setText(f"{perc:.2f} %")
+        self.lb_hw_view_level.setText(f"{perc:.2f}%, {charging_label}")
             
         self.notify_observers_with_value( (perc, is_charging) )
 
-        # update graph on Factory View
-        active_curve = self.ctl.multispec.get_hardware_feature_curve(self.name, spec.device_id)
-        if active_curve is None:
-            return
-        try:
-            x, y = rds.get_relative_to_now()
-            log.debug(f"updating battery charge curve ({len(y)} values)")
-            self.ctl.graph.set_data(curve=active_curve, y=y, x=x)
-        except:
-            log.error("error plotting battery data", exc_info=1)
-
-    def populate_placeholder(self):
-        log.debug(f"adding battery graph")
-        cfu = self.ctl.form.ui
-        cfu.battery_graph = pyqtgraph.PlotWidget(name="Battery Data")
-        cfu.battery_graph.invertX(True)
-        cfu.stackedWidget_battery.addWidget(cfu.battery_graph)
-        cfu.stackedWidget_battery.setCurrentIndex(1)
-
-        cfu.battery_graph.setMouseEnabled(x=False, y=False)
-
-    def add_spec_curve(self, spec):
-        if self.ctl.multispec.check_hardware_curve_present(self.name, spec.device_id):
-            return
-        curve = self.ctl.form.ui.battery_graph.plot([], pen=spec.curve.opts['pen'],name=str(spec.label))
-        self.ctl.multispec.register_hardware_feature_curve(self.name, spec.device_id, curve) 
-    
-    def remove_spec_curve(self, spec):
-        cur_curve = self.ctl.multispec.get_hardware_feature_curve(self.name, spec.device_id)
-        if cur_curve is None:
-            return
-        # remove current curve from graph
-        for curve in self.ctl.form.ui.battery_graph.listDataItems():
-            if curve.name() == cur_curve.name():
-                self.ctl.form.ui.battery_graph.removeItem(curve)
-        self.ctl.multispec.remove_hardware_curve(self.name, spec.device_id)
-        if self.ctl.multispec.get_spectrometers() == []:
-            self.lb_perc.setText("99.99 %")
-
-    def clear_data(self):
-        for spec in self.ctl.multispec.get_spectrometers():
-            log.debug(f"clearing battery history for {spec}")
-            if spec is None:
-                continue
-            app_state = spec.app_state
-            rds = app_state.battery_data
-            rds.clear()
-
-    def update_curve_color(self, spec):
-        curve = self.ctl.multispec.get_hardware_feature_curve(self.name, spec.device_id)
-        if curve is None:
-            return
-        curve.opts["pen"] = spec.color
-
-    def copy_data(self):
-        spec = self.ctl.multispec.current_spectrometer()
-        copy_str = []
-        for spec in self.ctl.multispec.get_spectrometers():
-            log.debug(f"copying battery history for {spec}")
-            if not self.ctl.multispec.check_hardware_curve_present(self.name, spec.device_id):
-                continue
-
-            app_state = spec.app_state
-            if app_state is None:
-                continue
-            rds = app_state.battery_data
-            copy_str.append(rds.get_csv_data("Battery Data",spec.label))
-        self.ctl.clipboard.raw_set_text('\n'.join(copy_str))
-
     def update_visibility(self):
         spec = self.ctl.multispec.current_spectrometer()
-        if spec is None:
-            visible = False
-        else:
-            if spec.settings.eeprom.has_battery:
-                visible = True
-            else:
-                visible = False
-
-        self.ctl.form.ui.frame_hardware_capture_battery.setVisible(visible)
+        visible = spec and spec.settings.eeprom.has_battery
+        self.strip_chart_level.set_visible(visible)
 
     def get_perc(self, spec=None):
-        if spec is None:
-            spec = self.ctl.multispec.current_spectrometer()
-        if spec is None:
-            return
-
-        if not spec.settings.eeprom.has_battery:
-            return
-
-        if spec.app_state.battery_data.empty():
-            return
-
-        (_, perc) = spec.app_state.battery_data.latest()
-        return perc
+        return self.strip_chart_level.get_latest()